requirements.txt
numpy>=1.19.5
Pillow>=8.1.2
requests>=2.25.1